made particle setForce function protected instead of private
added function to clear all the registered particle force generators

moved SIMULATION_SCALE from PhysicsDefines to PhysicsGlobals
disabled COMPLEX_POSITION

setup physics when ResetView is called so objects don't keep their velocities after resetting
clear the registered particle forces when setting up physics to prevent duplicate registrations (which would STACK!)
removed sphere batch for model1 as batch is set by Planet class

when GenerateGravity is called on a Planet, check if there is already a gravity force and return that if there is
added name to planet
added variables to store the initial distance from sun and initial speed so they can be used for reset
created a bunch of getter setter declarations for implied values
implemented copy constructor and equals operator

add PlanetaryData.ini to project for easy access

+++ libs/nalexanderhim/PhysicsObject.h
@@ -11,6 +11,8 @@ I certify that this assignment is entirely my own work.
 #include "Particle.h"
 #include "Object3D.h"
 
+#include "PhysicsGlobals.h"
+
 class PhysicsObject :
 	public Particle, public Object3D
 {
@@ -24,7 +26,7 @@ private:
 	}
 	inline void RefreshPhysicsPosition()//update the physics position based on the object position
 	{
-		mPosition = mLocalTransform.position * SIMULATION_SCALE;
+		mPosition = mLocalTransform.position * (float)SIMULATION_SCALE;
 	}
 public:
 	PhysicsObject(float initialMass)
